requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
python-decouple
django-crispy-forms
crispy-bootstrap5
django-filter
stripe
requests
gunicorn
